+++ deno.jsonc
@@ -32,11 +32,8 @@
     "build_npm": "deno run -A --no-check=remote _/tasks/build_npm.ts",
     "bootstrap": "deno task download_frame_metadata && deno task build_wasm && deno task star",
     "lock": "deno cache --lock=lock.json --lock-write --no-check=remote target/star.ts",
-    "example:balance": "deno run -A --no-check=remote examples/balance.ts",
-    "example:effect": "deno run -A --no-check=remote examples/effect.ts",
+    "example:read_balance": "deno run -A --no-check=remote examples/read_balance.ts",
     "example:playground": "deno run -A --no-check examples/playground.ts",
-    "example:ws": "deno run -A --no-check=remote rpc/test.ts",
-    "example:effect2": "deno run -A --no-check=remote effect/test.ts",
     "test": "deno test -A --no-check=remote -L=info",
     "update_snapshots": "deno test -A --no-check=remote -L=info -- --update"
   }
