+++ frame_metadata/Key.test.ts
@@ -1,4 +1,4 @@
-import { hashers } from "/runtime/Hashers.ts";
+import { hashers } from "/env/mod.ts";
 import * as asserts from "std/testing/asserts.ts";
 import { encodeKey } from "./Key.ts";
 import { accountId32, getLookupAndDeriveCodec } from "./test-util.ts";
